Replacement of Docling with Tesseract and PyMuPDF
Codigo do docling ta comentado caso alguem queira tentar refazer.
