+++ scripts/test_semantic_search.py
