+++ src/__init__.py
@@ -1,0 +1,2 @@
+"""Projeto Agente-IA-i9 (package root)."""
+
