+++ src/infrastructure/chroma_manager.py
@@ -1,7 +1,9 @@
+import os
+from typing import Any, Dict, List, Optional
+
 import chromadb
 from chromadb.config import Settings
-from typing import List, Dict, Any, Optional
-import os
+
 
 class ChromaManager:
     # gerencia as coleções do ChromaDB para o projeto.
@@ -12,22 +14,22 @@ class ChromaManager:
         self.persist_directory = persist_directory
 
         os.makedirs(persist_directory, exist_ok=True)
-        
+
         self.client = chromadb.PersistentClient(
             path=persist_directory,
             settings=Settings(anonymized_telemetry=False)
         )
-        
+
         self._collections = {}
-    
+
     def get_or_create_collection(self, name: str) -> chromadb.Collection:
-        
+
         # obtém ou cria uma coleção pelo nome.
 
         if name not in self._collections:
             self._collections[name] = self.client.get_or_create_collection(name=name)
         return self._collections[name]
-    
+
     def index_documents(
         self,
         collection_name: str,
@@ -36,7 +38,7 @@ class ChromaManager:
         metadatas: Optional[List[Dict[str, Any]]] = None
     ) -> None:
         # indexa documentos em uma coleção.
-        
+
         """Args:
             collection_name: 'editais' ou 'perfil_empresa'
             ids: Lista de IDs únicos para cada chunk
@@ -44,13 +46,13 @@ class ChromaManager:
             metadatas: Lista de metadados para cada chunk (tipo, data, fonte)
         """
         collection = self.get_or_create_collection(collection_name)
-        
+
         collection.add(
             ids=ids,
             documents=documents,
             metadatas=metadatas
         )
-    
+
     def get_collection_info(self, collection_name: str) -> dict:
         collection = self.get_or_create_collection(collection_name)
         return {
