+++ src/pipeline/__init__.py
@@ -1,0 +1,2 @@
+"""Pipeline package (Pipe & Filter)."""
+
