+++ src/pipeline/filters/__init__.py
@@ -1,0 +1,2 @@
+"""Filtros do pipeline."""
+
