+++ src/pipeline/ingestion/__init__.py
@@ -1,0 +1,2 @@
+"""Módulos de ingestão (PDF -> Markdown)."""
+
